+++ test/just-compile.c
@@ -1,0 +1,6 @@
+#!/usr/bin/env crun
+
+int main(void) {
+    puts("should not run");
+    return 0;
+}